Moves char_template_joiner out to its own header file
